Move remainin defult style and style functions to style module
